Decided to make a generic PadTest instead of MovePadTest, overloaded constructor in Player it can now take default and specific parameters, found a float/double bug in Pad and Player that is fixed now

+++ RoungOut/core/src/Objects/Pad.java
@@ -41,7 +41,7 @@ public class Pad {
         this.padSpeed = speed;
     }
 
-    public Pad(float length, float height, float padXPos, float padYPos,float padSpeed) {
+    public Pad(float length, float height, double padXPos, double padYPos,float padSpeed) {
         this.length = length;
         this.height = height;
         this.padXPos = padXPos;
